+++ MarioProject/PortSpriteIDs.h
@@ -1,5 +1,5 @@
 #pragma once
 
-constexpr int ID_SPRITE_SMALL_MARIO_MOVEMENT = 3e6 + 2e3 + 0;
-constexpr int ID_SPRITE_BIG_MARIO_MOVEMENT = 3e6 + 2e3 + 1;
-constexpr int ID_SPRITE_FLY_MARIO_MOVEMENT = 3e6 + 2e3 + 2;
+constexpr int ID_SPRITE_SMALL_MARIO_MOVEMENT = static_cast<int>(3e6 + 2e3 + 0);
+constexpr int ID_SPRITE_BIG_MARIO_MOVEMENT = static_cast<int>(3e6 + 2e3 + 1);
+constexpr int ID_SPRITE_FLY_MARIO_MOVEMENT = static_cast<int>(3e6 + 2e3 + 2);
